move util functions to separate module
